+++ src/lib/target-tasks/typedoc.ts
@@ -1,9 +1,7 @@
 import { Gulp } from "gulp";
 import gulpTypedoc from "gulp-typedoc";
-import { CompilerOptionsJson } from "../options/tsc";
 import { AbsPosixPath } from "../types";
 import { TaskFunction } from "../utils/gulp-task-function";
-import { deleteUndefinedProperties } from "../utils/utils";
 import { ResolvedTsLocations, resolveTsLocations, TypescriptConfig } from "./_typescript";
 
 export interface TypedocOptions {
@@ -13,22 +11,16 @@ export interface TypedocOptions {
 
 export function getTypedocTask(gulp: Gulp, tsConfig: TypescriptConfig, options: TypedocOptions): TaskFunction {
   const resolved: ResolvedTsLocations = resolveTsLocations(tsConfig);
-  const tscOptions: CompilerOptionsJson = {
-    ...tsConfig.tscOptions,
-    rootDir: resolved.rootDir,
-    outDir: resolved.outDir,
-    typeRoots: resolved.typeRoots,
-  };
-  deleteUndefinedProperties(tscOptions);
 
   const task: TaskFunction = function () {
     return gulp
       .src(resolved.absScripts, {base: tsConfig.srcDir})
-      // TODO(demurgos): Remove cast once rogierschouten/gulp-typedoc#24 is merged
+      // TODO(demurgos): Fix gulp typedoc: add missing `tsconfig` key?
       .pipe(gulpTypedoc(<any> {
         // TypeScript options (see typescript docs)
         module: "commonjs",
         target: "es2017",
+        tsconfig: resolved.tsconfigJson,
         includeDeclarations: false,
         out: options.dir,
         // TypeDoc options (see typedoc docs)